fix circular include by adding another header
